-Walmost-everything (#1148)
* start with -Weverything

* more

* some more

* bla

* bla

* bla

* bl

* bla

* bla

* Type casts for histograms

* bla

* bla

* blablub

* bla

* grrr

* test broken

* Revert "test broken"

This reverts commit ca3a8a55cf64d4d88438b231190635989ba903de.

* pgasus

* bla

* blub

* adawdgp

* Wasn jetzt noch?

* bla

* bla

* bla

* hast ja recht

* hmm

* what now?

* Revert "hmm"

This reverts commit f25637aa6c2688c1a78fa0c0ca91820ede6752fc.

* bla

* test

* So?

* bla

* fu

* aefawefawef

* Weiß ich doch

* Start fixing RangeFilter test

* Handle overflows.

* Improve performance by reusing distributions and more.

* Fix floating type problems by explicitely storing end values instead of calculating them.

* Add test to check for unsorted data throwing.

* format

* Update sql-parser

* bla

* bla

* Update src/lib/statistics/chunk_statistics/range_filter.hpp


* bla

* bla

* Remove unused argument

* bla

* dschit

* disable test

* bla

* bafawef

* constructor

* review

+++ src/test/statistics/chunk_statistics/counting_quotient_filter_test.cpp
@@ -53,19 +53,19 @@ T get_test_value(size_t run) {
 
 template <>
 int32_t get_test_value<int32_t>(size_t run) {
-  return 123457 + run;
+  return static_cast<int32_t>(123457 + run);
 }
 template <>
 int64_t get_test_value<int64_t>(size_t run) {
-  return 123457 + run;
+  return static_cast<int64_t>(123457 + run);
 }
 template <>
 float get_test_value<float>(size_t run) {
-  return 123457.0f + run;
+  return static_cast<float>(123457.0f + run);
 }
 template <>
 double get_test_value<double>(size_t run) {
-  return 123457.0 + run;
+  return static_cast<double>(123457.0 + run);
 }
 template <>
 std::string get_test_value<std::string>(size_t run) {
